Add/Enable DSA support for Linksys Armada devices
DSA has been fully supported in the kernel in commit 9f3f61a but
only for the Turris Omnia. This was due to the application of a previous
commit 7b33f6c that disabled it per DTS. Just change that file and other
necessary files to enable DSA for Linksys devices based on Armada 385.

v2: * Remove references to swconfig in cortex-a9.mk for Linksys devices
    * Remove 003-add_switch_nodes.patch removing dependencies on mvsw61xx

v3: * Expand this commit to affect all Linksys (385 + XP/375) boards
    * Remove mvsw61xx from config-4.14 as it depends on swconfig

v4: * Don't know if it will affect anything, but I reversed the order of
      ports in 02_network

v5: * User neheb suggested that I remove swconfig in the kconfig

v6: * Remove 075-dsa-inherit-parent-mac.patch as a test to see if it
      causes the issues with eth1; RGMII-ID interface from starting up

v7: * Fixup the LEDs on first boot

v8: * Modify how the MAC address is obtained